requirements.txt
keras==2.12.0
matplotlib==3.7.1
numpy==1.23.5
opencv_python==4.7.0.72
Pillow==9.5.0
tensorflow==2.12.0
tensorflow_intel==2.12.0
